backend/src/modules/products/services/UpdateProductService.ts
import AppError from '@shared/errors/AppError';
import { injectable, inject } from 'tsyringe';
import Product from '../infra/typeorm/entities/Product';
import IProductsRepository from '../repositories/IProductsRepository';

interface IRequest {
  id: string;
  name: string;
  description: string;
  price: number;
}

@injectable()
class UpdateProductService {
  constructor(
    @inject('ProductsRepository')
    private productsRepository: IProductsRepository,
  ) {}

  async execute({ id, name, description, price }: IRequest): Promise<Product> {
    const product = await this.productsRepository.findById(id);

    if (!product) {
      throw new AppError('Product does not exist');
    }

    Object.assign(product, {
      name,
      description,
      price,
    });

    await this.productsRepository.update(product);

    return product;
  }
}

export default UpdateProductService;
